Make parser emit an AST
Can now parse package, import and class definitions.  Class body (the bulk
of the language) still to be done.

Add the test.as file.  This file contains sample AS3 code that can be fed
to the parser.

Add a main function.  This function accepts a command line arg representing
the name of an AS3 file.   The file is opened and its contents fed to the
AS3 parser.  The output of the program is an ASCII tree representation of
AST created by the parser.